+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ## [0.0.2]
  * Add description to package.
  * Add example.
+
+ ## [0.1.0]
+ * Now you can change the color of list group
+ * borderColor, borderRadius and borderWidth are available as optional values for list group.